requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
SQLAlchemy==2.0.36
psycopg[binary]==3.2.3
passlib[bcrypt]==1.7.4
bcrypt==3.2.2
python-jose[cryptography]==3.3.0
pydantic==2.10.6
pydantic-settings==2.7.1
httpx==0.27.0
python-multipart==0.0.9
fpdf2==2.7.8
redis==5.0.1
rq==1.15.1
alembic==1.13.1
email-validator==2.3.0
beautifulsoup4==4.12.3
svglib>=1.6.0
reportlab>=4.4.10
pillow>=12.0.0
